+++ feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/component/CreateQuestionContent.kt
@@ -1,0 +1,65 @@
+package kr.boostcamp_2024.course.quiz.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import kr.boostcamp_2024.course.designsystem.ui.theme.component.WeQuizTextField
+import kr.boostcamp_2024.course.quiz.R
+
+@Composable
+fun CreateQuestionContent(
+    modifier: Modifier = Modifier,
+    focusRequester: FocusRequester,
+    keyboardOptions: KeyboardOptions,
+    keyboardActions: KeyboardActions,
+    title: String,
+    description: String? = null,
+    solution: String? = null,
+    onTitleChanged: (String) -> Unit,
+    onDescriptionChanged: (String) -> Unit,
+    onSolutionChanged: (String) -> Unit,
+) {
+
+    Column(
+        modifier = modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(10.dp),
+    ) {
+        WeQuizTextField(
+            modifier = Modifier.focusRequester(focusRequester),
+            label = stringResource(id = R.string.txt_question_title_label),
+            text = title,
+            onTextChanged = onTitleChanged,
+            placeholder = stringResource(id = R.string.txt_question_title_placeholder),
+            keyboardOptions = keyboardOptions,
+            keyboardActions = keyboardActions,
+        )
+        WeQuizTextField(
+            label = stringResource(id = R.string.txt_question_content_label),
+            text = description ?: "",
+            onTextChanged = onDescriptionChanged,
+            placeholder = stringResource(id = R.string.txt_question_content_placeholder),
+            minLines = 6,
+            maxLines = 6,
+            keyboardOptions = keyboardOptions,
+            keyboardActions = keyboardActions,
+        )
+        WeQuizTextField(
+            label = stringResource(id = R.string.txt_question_description_label),
+            text = solution ?: "",
+            onTextChanged = onSolutionChanged,
+            placeholder = stringResource(id = R.string.txt_question_description_placeholder),
+            minLines = 6,
+            maxLines = 6,
+            keyboardOptions = keyboardOptions,
+            keyboardActions = keyboardActions,
+        )
+    }
+}
